feat: 完善接口设计，完善文档内容
