--- app/src/main/java/com/origeek/pickerDemo/ui/component/Coil.kt
@@ -1,18 +1,0 @@
-package com.origeek.pickerDemo.ui.component
-
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.graphics.painter.Painter
-import androidx.compose.ui.platform.LocalContext
-import coil.compose.rememberAsyncImagePainter
-import coil.request.ImageRequest
-
-@Composable
-fun rememberCoilImagePainter(image: Any): Painter {
-    // 加载图片
-    val imageRequest = ImageRequest.Builder(LocalContext.current)
-        .data(image)
-        .size(coil.size.Size.ORIGINAL)
-        .build()
-    // 获取图片的初始大小
-    return rememberAsyncImagePainter(imageRequest)
-}
