--- app/src/main/res/drawable/ic_launcher_foreground.xml
@@ -1,32 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="108dp"
-    android:height="108dp"
-    android:viewportWidth="108"
-    android:viewportHeight="108">
-  <group android:scaleX="0.0421875"
-      android:scaleY="0.0421875"
-      android:translateX="32.4"
-      android:translateY="32.4">
-    <path
-        android:pathData="M81.92,296.96h61.44v61.44H81.92zM143.36,296.96h61.44v61.44H143.36zM204.8,296.96h61.44v61.44h-61.44zM266.24,296.96h61.44v61.44h-61.44zM327.68,296.96h61.44v61.44h-61.44zM327.68,235.52h61.44v61.44h-61.44zM389.12,174.08h61.44v61.44h-61.44zM389.12,112.64h61.44v61.44h-61.44z"
-        android:fillColor="#553A49"/>
-    <path
-        android:pathData="M450.56,112.64h61.44v61.44h-61.44zM389.12,235.52h61.44v61.44h-61.44zM512,112.64h61.44v61.44h-61.44z"
-        android:fillColor="#FDEC57"/>
-    <path
-        android:pathData="M450.56,174.08h61.44v61.44h-61.44zM512,174.08h61.44v61.44h-61.44zM450.56,235.52h61.44v61.44h-61.44zM512,235.52h61.44v61.44h-61.44zM389.12,296.96h61.44v61.44h-61.44zM450.56,296.96h61.44v61.44h-61.44zM512,296.96h61.44v61.44h-61.44zM573.44,296.96h61.44v61.44h-61.44zM389.12,358.4h61.44v61.44h-61.44zM450.56,358.4h61.44v61.44h-61.44zM512,358.4h61.44v61.44h-61.44zM573.44,358.4h61.44v61.44h-61.44zM81.92,419.84h61.44v61.44H81.92zM143.36,419.84h61.44v61.44H143.36zM204.8,419.84h61.44v61.44h-61.44zM266.24,419.84h61.44v61.44h-61.44zM327.68,419.84h61.44v61.44h-61.44zM389.12,419.84h61.44v61.44h-61.44zM450.56,419.84h61.44v61.44h-61.44zM512,419.84h61.44v61.44h-61.44zM573.44,419.84h61.44v61.44h-61.44zM634.88,419.84h61.44v61.44h-61.44zM696.32,419.84h61.44v61.44h-61.44zM757.76,419.84h61.44v61.44h-61.44zM819.2,419.84h61.44v61.44h-61.44zM880.64,419.84h61.44v61.44h-61.44zM143.36,481.28h61.44v61.44H143.36zM204.8,481.28h61.44v61.44h-61.44zM266.24,481.28h61.44v61.44h-61.44zM327.68,481.28h61.44v61.44h-61.44zM389.12,481.28h61.44v61.44h-61.44zM450.56,481.28h61.44v61.44h-61.44zM512,481.28h61.44v61.44h-61.44zM573.44,481.28h61.44v61.44h-61.44zM634.88,481.28h61.44v61.44h-61.44zM696.32,481.28h61.44v61.44h-61.44zM757.76,481.28h61.44v61.44h-61.44zM819.2,481.28h61.44v61.44h-61.44zM204.8,542.72h61.44v61.44h-61.44zM266.24,542.72h61.44v61.44h-61.44z"
-        android:fillColor="#FDD14A"/>
-    <path
-        android:pathData="M266.24,604.16h61.44v61.44h-61.44zM327.68,604.16h61.44v61.44h-61.44zM389.12,604.16h61.44v61.44h-61.44zM450.56,604.16h61.44v61.44h-61.44zM512,604.16h61.44v61.44h-61.44zM573.44,604.16h61.44v61.44h-61.44zM634.88,604.16h61.44v61.44h-61.44zM696.32,604.16h61.44v61.44h-61.44zM266.24,665.6h61.44v61.44h-61.44zM327.68,665.6h61.44v61.44h-61.44zM389.12,665.6h61.44v61.44h-61.44zM450.56,665.6h61.44v61.44h-61.44zM512,665.6h61.44v61.44h-61.44zM573.44,665.6h61.44v61.44h-61.44zM634.88,665.6h61.44v61.44h-61.44zM696.32,665.6h61.44v61.44h-61.44zM204.8,727.04h61.44v61.44h-61.44zM266.24,727.04h61.44v61.44h-61.44zM327.68,727.04h61.44v61.44h-61.44zM389.12,727.04h61.44v61.44h-61.44zM450.56,727.04h61.44v61.44h-61.44zM512,727.04h61.44v61.44h-61.44zM573.44,727.04h61.44v61.44h-61.44zM634.88,727.04h61.44v61.44h-61.44zM696.32,727.04h61.44v61.44h-61.44zM757.76,727.04h61.44v61.44h-61.44zM204.8,788.48h61.44v61.44h-61.44zM266.24,788.48h61.44v61.44h-61.44zM327.68,788.48h61.44v61.44h-61.44zM389.12,788.48h61.44v61.44h-61.44zM573.44,788.48h61.44v61.44h-61.44zM634.88,788.48h61.44v61.44h-61.44zM696.32,788.48h61.44v61.44h-61.44zM757.76,788.48h61.44v61.44h-61.44zM757.76,849.92h61.44v61.44h-61.44zM696.32,849.92h61.44v61.44h-61.44zM204.8,849.92h61.44v61.44h-61.44zM266.24,849.92h61.44v61.44h-61.44z"
-        android:fillColor="#FEBA3D"/>
-    <path
-        android:pathData="M327.68,542.72h61.44v61.44h-61.44zM389.12,542.72h61.44v61.44h-61.44zM450.56,542.72h61.44v61.44h-61.44zM512,542.72h61.44v61.44h-61.44zM573.44,542.72h61.44v61.44h-61.44zM634.88,542.72h61.44v61.44h-61.44zM696.32,542.72h61.44v61.44h-61.44zM757.76,542.72h61.44v61.44h-61.44z"
-        android:fillColor="#FDD14A"/>
-    <path
-        android:pathData="M573.44,235.52h61.44v61.44h-61.44zM634.88,358.4h61.44v61.44h-61.44zM327.68,358.4h61.44v61.44h-61.44zM266.24,358.4h61.44v61.44h-61.44zM204.8,358.4h61.44v61.44h-61.44zM143.36,358.4h61.44v61.44H143.36zM81.92,358.4h61.44v61.44H81.92zM696.32,358.4h61.44v61.44h-61.44zM757.76,358.4h61.44v61.44h-61.44zM819.2,358.4h61.44v61.44h-61.44zM880.64,358.4h61.44v61.44h-61.44z"
-        android:fillColor="#FDEC57"/>
-    <path
-        android:pathData="M450.56,51.2h61.44v61.44h-61.44zM512,51.2h61.44v61.44h-61.44zM573.44,112.64h61.44v61.44h-61.44zM573.44,174.08h61.44v61.44h-61.44zM634.88,235.52h61.44v61.44h-61.44zM634.88,296.96h61.44v61.44h-61.44zM696.32,296.96h61.44v61.44h-61.44zM757.76,296.96h61.44v61.44h-61.44zM819.2,296.96h61.44v61.44h-61.44zM880.64,296.96h61.44v61.44h-61.44zM942.08,358.4h61.44v61.44h-61.44zM942.08,419.84h61.44v61.44h-61.44zM880.64,481.28h61.44v61.44h-61.44zM819.2,542.72h61.44v61.44h-61.44zM757.76,604.16h61.44v61.44h-61.44zM757.76,665.6h61.44v61.44h-61.44zM204.8,665.6h61.44v61.44h-61.44zM204.8,604.16h61.44v61.44h-61.44zM143.36,542.72h61.44v61.44H143.36zM81.92,481.28h61.44v61.44H81.92zM20.48,419.84h61.44v61.44H20.48zM20.48,358.4h61.44v61.44H20.48zM819.2,727.04h61.44v61.44h-61.44zM819.2,788.48h61.44v61.44h-61.44zM819.2,849.92h61.44v61.44h-61.44zM819.2,911.36h61.44v61.44h-61.44zM143.36,727.04h61.44v61.44H143.36zM143.36,788.48h61.44v61.44H143.36zM143.36,849.92h61.44v61.44H143.36zM143.36,911.36h61.44v61.44H143.36zM757.76,911.36h61.44v61.44h-61.44zM696.32,911.36h61.44v61.44h-61.44zM634.88,849.92h61.44v61.44h-61.44zM573.44,849.92h61.44v61.44h-61.44zM512,788.48h61.44v61.44h-61.44zM450.56,788.48h61.44v61.44h-61.44zM389.12,849.92h61.44v61.44h-61.44zM327.68,849.92h61.44v61.44h-61.44zM266.24,911.36h61.44v61.44h-61.44zM204.8,911.36h61.44v61.44h-61.44z"
-        android:fillColor="#553A49"/>
-  </group>
-</vector>
