+++ sample/src/main/java/com/origeek/pickerDemo/ui/component/Layout.kt
@@ -1,0 +1,26 @@
+package com.origeek.pickerDemo.ui.component
+
+import androidx.compose.material.MaterialTheme
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.SideEffect
+import androidx.compose.ui.graphics.Color
+import com.google.accompanist.insets.ProvideWindowInsets
+import com.google.accompanist.systemuicontroller.rememberSystemUiController
+import com.origeek.pickerDemo.ui.theme.ImageViewerTheme
+
+@Composable
+fun BasePage(content: @Composable () -> Unit) {
+    ImageViewerTheme {
+        ProvideWindowInsets {
+            val systemUiController = rememberSystemUiController()
+            val useDarkIcons = MaterialTheme.colors.isLight
+            SideEffect {
+                systemUiController.setSystemBarsColor(
+                    color = Color.Transparent,
+                    darkIcons = useDarkIcons
+                )
+            }
+            content()
+        }
+    }
+}
