+++ sample/src/main/java/com/origeek/pickerDemo/ui/theme/Color.kt
@@ -1,0 +1,34 @@
+package com.origeek.pickerDemo.ui.theme
+
+import androidx.compose.ui.graphics.Color
+
+// light
+val light_f = Color(0xFFFFFFFF)
+
+// font
+val font_f = Color(0XFF111727)
+
+// accent
+val accent_f = Color(0xFF456DE6)
+val accent_dark_f = Color(0xFF999999)
+
+// second
+val second_f = Color(0xFFF3F5FD)
+
+// second
+val surface = Color(0x08456DE6)
+
+// error
+val error_f = Color(0xFFAC483B)
+
+// success
+val success_f = Color(0xFFB6D7C4)
+
+// back
+val back_dark_f = Color(0xFF272727)
+
+val pink_f = Color(0xFFF3ACC6)
+val royal_f = Color(0xFFBAA3FC)
+val stable_f = Color(0xFFEFEFEF)
+
+val dddd = Color(0xCCFFFFFF)
