+++ sample/src/main/res/values/colors.xml
@@ -1,0 +1,37 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <color name="transparent">#00FFFFFF</color>
+    <color name="black">#FF000000</color>
+
+    <!-- light -->
+    <color name="light_f">#FFFFFFFF</color>
+    <color name="light_m">#CCFFFFFF</color>
+    <color name="light_d">#99FFFFFF</color>
+    <color name="light_w">#00FFFFFF</color>
+    <color name="light_l">#33FFFFFF</color>
+    <!-- font -->
+    <color name="font_f">#FF111727</color>
+    <color name="font_d">#99111727</color>
+    <color name="font_l">#33111727</color>
+    <color name="font_m">#CC111727</color>
+    <color name="font_m_f">#FF414552</color>
+    <!-- accent -->
+    <color name="accent_f">#FF456DE6</color>
+    <color name="accent_m">#CC456DE6</color>
+    <color name="accentDark">#FF0043B3</color>
+    <color name="accentDarkDown">#FF000018</color>
+    <color name="accentLight">#FF819BFF</color>
+    <!-- second -->
+    <color name="second_f">#FFAAB6CB</color>
+    <color name="secondDark">#FF7B869A</color>
+    <color name="secondLight">#FFDCE8FE</color>
+
+    <color name="error_f">#FFAC483B</color>
+    <color name="warn_f">#FFFFD9A0</color>
+    <color name="success_f">#FFB6D7C4</color>
+    <color name="pink_f">#FFF3ACC6</color>
+    <color name="royal_f">#FFBAA3FC</color>
+
+    <color name="stable_f">#FFEFEFEF</color>
+    <color name="stable_m">#CCEFEFEF</color>
+</resources>
