+++ settings.gradle
@@ -14,5 +14,5 @@ dependencyResolutionManagement {
     }
 }
 rootProject.name = "ImagePicker"
-include ':app'
+include ':sample'
 include ':imagePicker'
